move open generate ui logic to service
